Create README.md
This study explores the issue of automated control strategies. In the process of automated production, due tothe continuous change of environmental factors and hardware factors, the automatic control of fixed parameters often cannot achieve the accuracy and efficiency of production.Using the Pololu 3pi+ 32U4 as an example, we investigated how to maintain the stability and output power of the power system by adjusting the PID value of the motor output in real time. By using a genetic algorithm (GA) to update the PID compensation in real time, we compared three control strategies: the traditional bang-bang controller, fixed PID
compensation, and the method of adjusting the PID using a GA. Experiments were conducted to assess the straightline traveling ability and PWM compliance by measuring the encoder counts and wheel speeds of the trolley under different control strategies.